refactor CLI structure: move main app logic to commands.py, update entry points, remove unused files
