+++ src/cli/__init__.py
@@ -1,2 +1,6 @@
 # src/cli/__init__.py
 # CLI components module for Loom
+
+from .commands import app
+
+__all__ = ["app"]
