+++ src/main.py
@@ -1,7 +1,7 @@
 # src/main.py
 # Entry point for the Loom CLI application
 
-from .main_cli import app
+from .cli.commands import app
 
 if __name__ == "__main__":
     app()
